+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - added support for using custom and extension commands in saved prompts
 - added modelCallSettings property to onAgentStarted event for overriding AI SDK model call parameters
 - correctly handling multiple reasoning message parts
+- falling back to default agent profile when opening new project with agent profile that does not exist
 
 ## [0.75.0]